SPENDING-2  Сделал получение категорий, добавил состояние к пользователю

+++ src/main/java/ru/voitel/telegram/category/CategoryHandler.java
@@ -16,5 +16,5 @@ public interface CategoryHandler {
      * Получение категорий расходов пользователей
      * @param UserID id пользователя
      */
-    String getUserCategories(Long UserID);
+    String getUserCategories(Integer UserID);
 }
